Move openshiftImageSource into a separate openshift/openshift_src.go
We'd like to have easy access to both internal/imagesource/impl and
internal/imagedestination/impl, which is easier with separate files.

Only moves unchanged code, should not change behavior.
